add dev proxy and some renice

+++ inventory/src/main/java/com/github/thomasandre84/model/Item.java
@@ -1,7 +1,9 @@
 package com.github.thomasandre84.model;
 
+import java.math.BigDecimal;
+
 public class Item {
     String name;
-    Double value;
+    BigDecimal value;
     Room room;
 }
